pyproject.toml
[project]
name = "lotusql"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "notebook (>=7.4.5,<8.0.0)",
    "ipykernel (>=6.30.1,<7.0.0)",
    "google-adk (>=1.12.0,<2.0.0)",
    "streamlit (>=1.48.1,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "requests (>=2.32.5,<3.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "litellm (>=1.75.9,<2.0.0)"
]

[tool.poetry]
packages = [{include = "lotusql", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
